requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
bcrypt==4.3.0
certifi==2025.4.26
cffi==1.17.1
colorama==0.4.6
contextlib2==21.6.0
coverage==7.6.12
cryptography==44.0.2
Django==5.1.6
django-cors-headers==4.3.1
django-debug-toolbar==5.1.0
django-environ==0.12.0
django-ninja==1.3.0
django-ninja-extra==0.22.3
django-ninja-jwt==5.3.5
django-storages==1.14.6
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
iniconfig==2.0.0
injector==0.22.0
packaging==24.2
paramiko==3.5.1
pillow==11.1.0
pluggy==1.5.0
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.9.0
PyNaCl==1.5.0
pytest==8.3.4
pytest-cov==6.0.0
pytest-django==4.10.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
sentry-sdk==2.27.0
six==1.17.0
sqlparse==0.5.3
storage==0.0.4.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.4.0
